requirements.txt
pyyaml==6.0.2
python-dateutil==2.9.0.post0
playwright==1.49.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
pypdf
google-genai>=1.0.0
flask>=3.0.0
apscheduler>=3.10.0
gunicorn>=21.0.0
sendgrid>=6.11.0
markdown>=3.5.0

